refactor(svelte): make GitHub a per-project inline view, move its settings to Settings
Replace the full-screen GitHub overlay with an inline view opened per project
from each project card's ⋯ menu (GitHub → Pull Requests / Issues / Actions). It
takes the place of the center + right panels while the left sidebar and the
browser panel stay visible, is scoped to the opened project (no repo selector),
and shows only the chosen section. Its close (left), a section switcher, and
refresh (right) live in the view's own toolbar rather than a window-height
header. Overview is removed; the account/session panel and the GitHub
preferences move to Settings → GitHub.

Activating any worktree returns to the workspace and opens its terminal as
usual, so the inline view closes. This is centralized at the terminal/agent open
choke points (app.openTerminal / app.launchAgent) plus setActiveWorktree, so the
sidebar +, quick commands, agent launch and PR→worktree all leave the view
instead of opening a terminal hidden behind it.

- state: app.githubInline / openGithubInline / closeGithub; GithubSection is now
  pulls | issues | actions; SettingsSection gains "github"
- svelte: GitHub.svelte rewritten to the inline model (no nav sidebar, toolbar
  close/switcher/refresh, repo re-sync on open/repo-change/sign-in);
  GithubSettings.svelte (new) hosts account/session + prefs + AI PR authoring,
  rendered by Settings.svelte in the Workspace group above Updates; ProjectCard
  gains the GitHub submenu; GithubPanel "big view" buttons open the inline view;
  GithubStatusButton is now a passive indicator; SidebarProfile drops its GitHub
  item; +page.svelte renders the inline view in place of center+right
- keys: remove the openGitHub action / Ctrl+G binding
- i18n: settings.github + github.settings.desc; drop the orphaned overview / nav
  / shortcut keys
- docs: docs/github.md, architecture/02a §3 + 02c §6 + 04, CHANGELOG
