+++ uxnandesktop/src/lib/components/GithubStatusButton.svelte
@@ -1,11 +1,12 @@
 <script lang="ts">
-  // Status-bar GitHub button: opens the GitHub section, shows the unread
-  // notifications count (when enabled) and the API rate-limit remaining in a
-  // tooltip. Hidden when disabled or not signed in (nothing to show).
+  // Status-bar GitHub indicator (passive): shows the unread notifications count
+  // (when enabled) and the API rate-limit remaining in a tooltip. GitHub itself
+  // opens per-project from each project card's ⋯ menu, so this no longer
+  // navigates anywhere — it's purely informational. Hidden when disabled or not
+  // signed in (nothing to show).
   import { app } from "$lib/state/app.svelte";
   import { github } from "$lib/state/github.svelte";
   import { i18n } from "$lib/i18n";
-  import { cn } from "$lib/utils";
   import { TooltipSimple } from "$lib/components/ui/tooltip";
   import GitPullRequestIcon from "@lucide/svelte/icons/git-pull-request";
 
@@ -16,18 +17,17 @@
           remaining: github.rateLimit.remaining,
           limit: github.rateLimit.limit,
         })
-      : i18n.t("github.open"),
+      : i18n.t("github.title"),
   );
 </script>
 
 {#if show}
   <TooltipSimple title={tip}>
     {#snippet children(props)}
-      <button
+      <span
         {...props}
-        class="inline-flex items-center gap-1 rounded px-1 text-muted-foreground hover:text-foreground"
-        aria-label={i18n.t("github.open")}
-        onclick={() => app.openGitHub()}
+        class="inline-flex items-center gap-1 px-1 text-muted-foreground"
+        aria-label={i18n.t("github.title")}
       >
         <GitPullRequestIcon class="size-3.5" />
         {#if github.notifications > 0}
@@ -35,7 +35,7 @@
             {github.notifications}
           </span>
         {/if}
-      </button>
+      </span>
     {/snippet}
   </TooltipSimple>
 {/if}
