+++ uxnandesktop/src/lib/components/SidebarProfile.svelte
@@ -1,12 +1,10 @@
 <script lang="ts">
   // Left-sidebar footer: a configurable profile card (avatar + name + a line of
-  // text), shadcn-sidebar-footer style. Clicking it opens a menu with the
-  // GitHub and Settings sections (moved here from the quick-actions row) plus an
-  // "Edit profile" entry. The avatar/name/description live in
+  // text), shadcn-sidebar-footer style. Clicking it opens a menu with Settings
+  // plus an "Edit profile" entry. The avatar/name/description live in
   // `AppSettings.profile` and are edited via SidebarProfileDialog.
   import * as DropdownMenu from "$lib/components/ui/dropdown-menu";
   import { app } from "$lib/state/app.svelte";
-  import { github } from "$lib/state/github.svelte";
   import { cn } from "$lib/utils";
   import { icon, text, divider, row, focus } from "$lib/design";
   import { i18n } from "$lib/i18n";
@@ -17,15 +15,13 @@
   import UserRoundIcon from "@lucide/svelte/icons/user-round";
   import ChevronsUpDownIcon from "@lucide/svelte/icons/chevrons-up-down";
   import SettingsIcon from "@lucide/svelte/icons/settings";
-  import GithubIcon from "@lucide/svelte/icons/git-pull-request";
   import PencilIcon from "@lucide/svelte/icons/pencil";
 
   let editOpen = $state(false);
 
-  // Shortcut hints (the actions still fire from the global keybindings; these are
-  // just discoverability cues in the menu).
+  // Shortcut hint (the action still fires from the global keybindings; this is
+  // just a discoverability cue in the menu).
   const settingsBinding = $derived(resolveBinding("openSettings"));
-  const githubBinding = $derived(resolveBinding("openGitHub"));
 
   // Display fields with graceful fallbacks so the card always reads as intentional
   // (the user personalizes them from the Edit dialog).
@@ -60,11 +56,6 @@
             class="relative flex size-8 shrink-0 items-center justify-center rounded-md border border-sidebar-border/60 bg-sidebar-foreground/[0.04]"
           >
             <EntityIcon value={app.sidebarProfile.icon} class="size-5" fallback={avatarGlyph} />
-            {#if github.notifications > 0}
-              <span
-                class="absolute -right-0.5 -top-0.5 size-2 rounded-full bg-primary ring-2 ring-sidebar"
-              ></span>
-            {/if}
           </span>
           <span class="flex min-w-0 flex-1 flex-col">
             <span class={cn("truncate text-sidebar-foreground", text.bodyStrong)}>
@@ -100,19 +91,6 @@
         </span>
       </div>
       <DropdownMenu.Separator />
-      <DropdownMenu.Item class={cn(text.menu, "gap-2")} onclick={() => app.openGitHub()}>
-        <GithubIcon class={icon.button} />
-        <span class="flex-1">{i18n.t("github.title")}</span>
-        {#if github.notifications > 0}
-          <span
-            class="inline-flex min-w-4 items-center justify-center rounded-full bg-primary px-1 text-[10px] font-semibold text-primary-foreground"
-          >
-            {github.notifications}
-          </span>
-        {:else if githubBinding}
-          <KeyChord chord={githubBinding} />
-        {/if}
-      </DropdownMenu.Item>
       <DropdownMenu.Item class={cn(text.menu, "gap-2")} onclick={() => app.openSettings()}>
         <SettingsIcon class={icon.button} />
         <span class="flex-1">{i18n.t("settings.title")}</span>
