+++ uxnandesktop/src/lib/keybindings.ts
@@ -55,13 +55,6 @@ export const KEY_ACTIONS: KeyAction[] = [
     default: "Mod+,",
   },
   {
-    id: "openGitHub",
-    labelKey: "shortcuts.openGitHub",
-    descKey: "shortcuts.openGitHubDesc",
-    category: "general",
-    default: "Mod+G",
-  },
-  {
     id: "openQuickCommands",
     labelKey: "shortcuts.openQuickCommands",
     descKey: "shortcuts.openQuickCommandsDesc",
